ZmqSocket>>dataAvailable does not block when reading data.
Previously it was possible for global poller signal that there is data to be read, but for some reason somebody could have read that data before (or image restart might have occurred) and there is no actual data to be read. Trying to receiving data that was not there jammed the whole poller thread.